\name{getDefaultMappings}
\alias{getDefaultMappings}
\alias{getDefaultScalings}
\title{Get the default mappings and scalings for a given sonlayer shape}
\usage{
  getDefaultMappings(shape)

  getDefaultScalings(shape)
}
\arguments{
  \item{shape}{A character containing the name of the shape
  to be looked up}
}
\value{
  The default mappings or default scalings for a given
  shape.
}
\description{
  Given a \code{shape} name, these functions returns the
  default mappings or scalings for that shape.
}
\keyword{internal}

