\name{.getDefaultSettings}
\alias{.getDefaultSettings}
\title{Get the default settings and scalings for a given sonlayer shape}
\usage{
  .getDefaultSettings(shape)
}
\arguments{
  \item{shape}{A character containing the name of the shape
  to be looked up}
}
\value{
  The default settings or default scalings for a given
  shape.
}
\description{
  Given a \code{shape} name, these functions returns the
  default settings or scalings for that shape.
}
\keyword{internal}

