\name{internalsonify}
\alias{.checkData}
\alias{.checkRendering}
\alias{internalsonify}
\alias{rendering}
\title{Internal functions related to layering and rendering}
\usage{
  .checkRendering(x)

  .checkData(x)

  rendering(x)
}
\description{
  These all will eventually will be used as arguments to
  \code{sonify()} and represent layering functionality that
  is not yet actually implemented.
}
\keyword{internal}

