\name{shapeDescriptions}
\alias{shapeDescriptions}
\title{Show the names and descriptions of all available shapes}
\usage{
  shapeDescriptions()
}
\value{
  A \code{data.frame} with each row representing one of the
  shapes available for sonification, a brief description of
  the shape, and the parameters available for sonification.
}
\description{
  A simple function to show all the currently available
  shapes, descriptions, and available sonic parameters for
  mapping data onto.
}

