\name{sonsave}
\alias{sonsave}
\title{Save sonifications to an audio file}
\usage{
  sonsave(what, where, ...)
}
\arguments{
  \item{what}{A sonify object}

  \item{where}{The path to the desired \code{.wav} output}

  \item{\dots}{other arguments to be passed onto the
  rendering}
}
\description{
  Convenience function to save a \code{sonify} object to an
  audio (\code{.wav}) file
}

