\name{internaldf}
\alias{.dfNotes}
\alias{.getMappings}
\alias{.getData}
\alias{.getNotes}
\title{Internal functions to generate output \code{data.frame} from a
  \code{sonify} object. This is then used by whatever method renders the
  sounds.}
\usage{
.dfNotes(s)
.getMappings(x, sonlayernum)
.getData(x, sonlayernum)
.getNotes(x, sonlayernum)
}

\keyword{internal}
  