\name{render.csound}
\alias{render.csound}
\title{Rendering sound using csound}
\usage{
  render.csound(x, audioSample = FALSE, ...)
}
\arguments{
  \item{x}{A \code{score} object created by
  \code{\link{.getScore}}}

  \item{audioSample}{Currently ignored.}

  \item{\dots}{Currently ignored.}
}
\description{
  \code{render.csound} uses the csound package to render a
  \code{score} object.
}

